requirements.txt
openai
python-dotenv
rich
mcp
tiktoken